+++ src/AtomUI.Desktop.Controls/Skeleton/SkeletonAvatar.cs
@@ -5,30 +5,30 @@ using Avalonia.Data;
 
 namespace AtomUI.Desktop.Controls;
 
-public class SkeletonAvatar : AbstractSkeleton, ICustomizableSizeTypeAware
+public partial class SkeletonAvatar : AbstractSkeleton, ICustomizableSizeTypeAware
 {
     #region 公共属性定义
     public static readonly StyledProperty<AvatarShape> ShapeProperty =
         AvaloniaProperty.Register<SkeletonAvatar, AvatarShape>(nameof(Shape), AvatarShape.Circle);
-    
+
     public static readonly StyledProperty<CustomizableSizeType> SizeTypeProperty =
         CustomizableSizeTypeControlProperty.SizeTypeProperty.AddOwner<SkeletonAvatar>();
-    
+
     public static readonly StyledProperty<double> SizeProperty =
         AvaloniaProperty.Register<SkeletonAvatar, double>(nameof(Size), Double.NaN);
-    
+
     public AvatarShape Shape
     {
         get => GetValue(ShapeProperty);
         set => SetValue(ShapeProperty, value);
     }
-    
+
     public CustomizableSizeType SizeType
     {
         get => GetValue(SizeTypeProperty);
         set => SetValue(SizeTypeProperty, value);
     }
-    
+
     public double Size
     {
         get => GetValue(SizeProperty);
@@ -36,15 +36,15 @@ public class SkeletonAvatar : AbstractSkeleton, ICustomizableSizeTypeAware
     }
 
     #endregion
-    
+
     #region 内部属性定义
-    
+
     internal static readonly DirectProperty<SkeletonAvatar, bool> IsCustomSizeProperty =
         AvaloniaProperty.RegisterDirect<SkeletonAvatar, bool>(
             nameof(IsCustomSize),
             o => o.IsCustomSize,
             (o, v) => o.IsCustomSize = v);
-    
+
     private bool _isCustomSize;
 
     internal bool IsCustomSize
@@ -54,11 +54,11 @@ public class SkeletonAvatar : AbstractSkeleton, ICustomizableSizeTypeAware
     }
 
     #endregion
-    
+
     public SkeletonAvatar()
     {
     }
-    
+
     protected override void OnPropertyChanged(AvaloniaPropertyChangedEventArgs change)
     {
         base.OnPropertyChanged(change);
@@ -67,22 +67,23 @@ public class SkeletonAvatar : AbstractSkeleton, ICustomizableSizeTypeAware
             IsCustomSize = !double.IsNaN(Size);
             ConfigureSize();
         }
-        if (change.Property == ShapeProperty || 
-            change.Property == SizeTypeProperty || 
+        if (change.Property == ShapeProperty ||
+            change.Property == SizeTypeProperty ||
             change.Property == SizeProperty)
         {
             ConfigureShape();
         }
     }
-    
+
     private void ConfigureShape()
     {
         if (Shape == AvatarShape.Circle)
         {
-            SetValue(CornerRadiusProperty, new CornerRadius(Width), BindingPriority.Template);
+            var radius = double.IsNaN(Width) ? Height : Width;
+            SetValue(CornerRadiusProperty, new CornerRadius(radius), BindingPriority.Template);
         }
     }
-    
+
     private void ConfigureSize()
     {
         if (!double.IsNaN(Size))
@@ -92,7 +93,7 @@ public class SkeletonAvatar : AbstractSkeleton, ICustomizableSizeTypeAware
             SetValue(HeightProperty, Size, BindingPriority.Template);
         }
     }
-    
+
     protected override void OnApplyTemplate(TemplateAppliedEventArgs e)
     {
         base.OnApplyTemplate(e);